requirements.txt
openai>=1.0.0
jinja2>=3.1.0
weasyprint>=60.0
python-dotenv>=1.0.0
python-docx>=1.0.0
pypdf>=4.0.0
aiogram>=3.0.0

# Тестирование
pytest>=7.0.0
pytest-cov>=4.0.0
